refatoração do código realizada, no intuito de organizar melhor e disponibilizar dnd
