+++ src/App.jsx
@@ -1,5 +1,5 @@
 import LoginPage from './pages/LoginPage';
-import InitialPage from './pages/InitialPage';
+import Home from './pages/Home';
 import RecoverPage from './pages/RecoverPage';
 import CreateAccountPage from './pages/CreateAccountPage';
 import { Routes, Route } from 'react-router-dom';
@@ -11,7 +11,7 @@ function App () {
     <>
       <Routes>
         <Route path="/" element={<LoginPage />} />
-        <Route path="/home" element={<InitialPage />} />
+        <Route path="/home" element={<Home />} />
         <Route path="/recover" element={<RecoverPage />} />
         <Route path="/create" element={<CreateAccountPage />} />
         <Route path="*" element={<h1>404, not found </h1>} />
